+++ app/Http/Controllers/SkillController.php
@@ -10,12 +10,21 @@ use illuminate\Support\Str;
 class SkillController extends Controller
 {
     function skillView(){
-        return view('backend.pages.skills.view-skills',[
-            'skills' => Skill::paginate(10),
-        ]);
+
+        if(auth()->user()->can('view skill')){
+            return view('backend.pages.skills.view-skills',[
+                'skills' => Skill::paginate(10),
+            ]);
+        }else{
+            return abort('404');
+        }
     }
     function skillAdd(){
-        return view('backend.pages.skills.add-skill');
+        if(auth()->user()->can('add skill')){
+            return view('backend.pages.skills.add-skill');
+        }else{
+            return abort('404');
+        }
     }
     function skillUpdate(Request $request){
         $request->validate([
@@ -30,20 +39,26 @@ class SkillController extends Controller
         return redirect()->route('skillView')->with('success','Skill added.');
     }
     function skillDelete($id){
-        $skill = Skill::find($id);
-
-        if($skill){
-            $skill->delete();
-            return back()->with('success',' Skill deleted.');
+        if(auth()->user()->can('delete  management')){
+            $skill = Skill::find($id);
+            if($skill){
+                $skill->delete();
+                return back()->with('success',' Skill deleted.');
+            }else{
+                return back()->with('fail',' Skill delete failed.');
+            }
         }else{
-            return back()->with('fail',' Skill delete failed.');
-
+            return abort('404');
         }
     }
     function skillEdit($id){
-        return view('backend.pages.skills.edit-skill',[
-            'skill' =>Skill::findOrFail($id),
-        ]);
+        if(auth()->user()->can('edit skill')){
+            return view('backend.pages.skills.edit-skill',[
+                'skill' =>Skill::findOrFail($id),
+            ]);
+        }else{
+            return abort('404');
+        }
     }
     function skillEditUpdate(Request $request){
         $skillUpdate = Skill::findOrFail($request->id);
